+++ core/src/routes/(editor)/editor/+page.svelte
@@ -165,6 +165,7 @@
     
 <svelte:component this={Selector} />
 
+
 <style>
 
     #editorInnerPanel{